Add large level support

+++ src/xrCore/FS.cpp
@@ -465,7 +465,7 @@ CVirtualFileReader::CVirtualFileReader(const char* cFileName)
 {
 	// Open the file
 	hSrcFile = Platform::CreateFile(cFileName, false);
-	Size = (int)Platform::GetFileSize(hSrcFile);
+	Size = Platform::GetFileSize(hSrcFile);
 
 	if (Size == 0)
 		return;
